add ut and move codegen_feature_column package to codegen
