+++ pkg/sql/codegen/codegen_feature_column.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package feature
+package codegen
 
 import (
 	"fmt"
